add logs button on form window with moved init func
