+++ src/lib/stores.ts
@@ -1,5 +1,6 @@
 import { writable } from 'svelte/store';
-import type { FlightData } from '$lib/structs';
+import type { FlightData, FlightLog } from '$lib/structs';
 
 export const InfoVisible = writable(true);
 export const FlightRecord = writable({ initialized: false } as FlightData);
+export const LogArray = writable([] as FlightLog[]);
